requirements.txt
dnstwist==20230413
PyYAML==6.0
